refactor: hooks to open positions
